requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
anyio==3.6.2
APScheduler==3.10.1
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
autoflake==2.1.1
backcall==0.2.0
black==23.3.0
boto3==1.26.123
botocore==1.29.123
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.95.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
ipykernel==6.22.0
ipython==8.13.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.2.0
jupyter_core==5.3.0
langchain==0.0.147
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.24.3
openai==0.27.4
openapi-schema-pydantic==1.2.4
orjson==3.8.10
packaging==23.1
pandas==2.0.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.1
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
ruff==0.0.262
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
stack-data==0.6.2
starlette==0.26.1
tabulate==0.9.0
tenacity==8.2.2
toml==0.10.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
types-PyMySQL==1.0.19.6
types-requests==2.28.11.17
types-urllib3==1.26.25.10
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
vulture==2.7
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.2
yarl==1.9.1
